Developped the Post Trip Request action, and more:
UI:
- Implemented the UI dialog for the postTripRequest user action.
- Added new methods in the Input util class that input a certain input
type (i.e int, string etc.) and validate it is indeed of that type.
- Surrounded the main dialog loop with a general try-catch block that
prints stack traces as a last resort in order to preventing the program
to crash.

Engine:
- Implemented the postTripRequest() method.
- Implemented a getAllStations() method.
- Made the TranspoolManager a Singleton.
- Added a Users Set to the TranspoolManager, and a set of methods to
manipulate that coolection of users, for example: createUser() which
creates a user, adds it to the Set if there are no identical users
already within it, and returns the newly created user.
- Added a serialize() method to the TranspoolXmlSerializer.
- Fixed the bug of the xml loading that ignored the fact that roads may
be two way.
- Added a class that supposed to validate input from the user in the
backend, for example for the existance of certain stations (by name).
Removed the Map reference held by the TranspoolManager, since there is
no need to keep it in a variable. Thought, there still is a getMap()
method, that returns the getInstance() of the Map Singleton.
- Made the Sets of TripOffers/Requests in the TripsManager class
non-static, and implemented the addRequest() method.

Model:
- Fixed the equals() & hashCode() of User.
- Added a reset() method to the Map in order to allow users to load new
xml files that override the existing data in the system, while
maintaining the existance of only one Map object in the system at a
given moment.
- Created a new Exception that indicates invalid input from the user.
- Created a new Exception that indicates the creation a user that
already exist.
- Fully created the TripRequest class.

In Addition, fixed the Class-Path in the engine module's manifest file.