Almost finished backend for getTripOffers command
UI:
- Main: removed garbage code that was put there for testing.

Model:
- Added an Enum to indicate the type of transition a user goes through
in their src/dst stations (Departing, Arriving).
- Road: After checking the xml schema and realizing the types of the
length, gasNeededPerKm and maxDrivingSpeed are ints, I changed them from
doubles to ints.
In Addition, changed the calcRoadTravelDuration to it returns the number
of minutes the ride would take (int) instead of the duration in hours
(which would be double).

-TripOffer: Added an id member that generated automatically when a new
instance is created. Refactored the name of passengersQuota to
basePassengersCapacity.
Added functions that would be utilized when calling the getTripOffers
command.

TripRequest: Added an id member (same as in mechanism as in TripOffer).
Replaced the member references from station names to actual Station
objects. Added an isMatched member.

TripTiming: Generated getters & setters for the class members.
Added a method that rounds the number of minutes to the closest product
of 5.

User: Added an id member (You get the point..).
Added static methods that allows to get a user from a Collection by the
id of the wanted user.