Finished General Graph structure
Graph:
Removed the adjacency list because it was unnecessary and kept redundant
duplicates of data. The structure that each vertex keeps a set of all
its outwards edges is enough.
I won't make it to replace the currect implmentation with the graph
implementation on time, so it'll be kept for the next version.

Main:
Removed graph's tests.

Road:
Starting to implement the methods that find a path between stations.