Finished version 1 dev w/o bonuses. Fixing remaining bugs.
TranspoolXmlLoader:
Had a bug where I forgot to call getName() on a station so the default
toString() was called to mension the station.

Engine:
-Implemented the getAllMatchesToRequest() method.
It returns all possible matches for a given TripRequest.
-Implemented the matchTripRequestToOffer() method which registers the
given request to the given offer.

TripsManager:
-Added a functions that search for matches given a request, and a
function that registers a request to an offer (and the offer to the
request accordinglly).
- Added a method to return all unmatched trip requests.

IEngine:
Added the signatures for getAllMatchedToRequest() &
matchTripRequestToOffer()

Road:
-Removed commented-out code,
-Fixed the getSubPath() method.

TripOffer:
-Added a method that calculates the arrival time to a specified station.
-Added a method that registers a request to the registeredRequests set.
-Added a getTripDuration overload that receives a Station and returns
the amount of minutes it would take to get there.

TripRequest:
-Added static versions of some methods in order of these action would be
able to be used outside the context of a specific TripRequest.

DialogManager:
-Extracted the set of actions that print the details of a trip request
to a new method.
-Fixed some hardcoded strings.
-Added a method to get the unmatched trip requests.
-Added a method to print details of matches to requests.
-Refactored some methods to use the improved input overloads that
receive a string to prompt before waiting for the user to input.
-Added the main dialog function to start the matching process of
requests. It is written like shit. I'm not proud of it.

TranspoolConsole:
-Extracted the switch to a method in order to easily encapsulate it
within a try-catch block in order for the code to continue running if an
unexpected exception gets thrown.

Input:
As mentioned before, created reallo cool overloads for the input system.
Now one can use the getNonEnptyStringInput() and the getIntInput()
methods just like in python, where you give the method what to prompt to
the console before waiting to an input.

Output:
Removed a redundant point in a string.