Finished bonus 1
Engine:
-Added a getAllRoads() method.
-Added a psotTripOffer method.
-Added a method to get all unmatched requests.

TranspoolManager:
-Added a method that checks if a user exists and returns it. if it
doesn't exist it'll be created.

TripsManager:
-Modified the getRequests() method so now it returns a list of requests
(instead of a set) which are ordered by their ID.
- Fixed a bug in the getUnmatchedRequests which caused it to return only
matched requests.

RoadDoesNotExistException:
-Added a Ctor that receives the name of a source station and the name of
a destination station and formats the exception message to specify them
accordingly.

RepetitionRate:
-Added methods to get the number of elements in the enum and check if an
int value can be transformed to an enum value by checking if it's in the
range.

UserAction:
-Added the POST_TRIP_OFFER user action.

IEngine:
-Added the signatures for the getAllRoads(), postTripOffer() and
getUnmatchedTripRequests() methods.

-Map:
-Made the getStation() method throw a StationDoesNotExistException.
-Added a getStationIfExists() method that checks the station name
received and returns it if it exists. It returns null otherwise.
-Added a getStationsByNames() method that receives a list of station
names and returns the list of matching stations. Throws
StationDoesNotExistException if one of the stations received do not
exist.
-Made the getRoad() method throw a RoadDoesNotExistException.
-Added a getRoadsByStationsList() method that receives a list of
stations and returns the list of roads between them. Throws
RoadDoesNotExistException if the list of stations represent at least one
road that doesn't exist.

Road:
-Made the getRoadBySrcDst() method to throw a RoadDoesNotExistException.

Station:
-Made the getStationByName() method to throw a
StationDoesNotExistException.
-Commented-out the getStationsInPath() method since it is not used
anywhere and I'm not sure if I need it.

TripRequest:
-Made the Ctor use the Map.getStationIfExist() method in order for it to
assign null if the station doesn't exist instead of throwing an
exception.
-Fixed a bug in the static getArrivalTime() method that distored the
arrival time by using the start time of the trip offer instead of the
start time of the trip request.

DialogManager:
-Added a method to print all roads.
-Removed a redundant variable in getAllTripRequests() method.
-Added a method that sorts IdentifiableTranspoolEntity collections by
the IDs of the items.
-Made the getUnmatchedTripRequests() method use the engine's
getUnmatchedTripRequests() method. That line contains a Warning of an
unchecked cast' though its OK.
-Added the postTripOfferDialog() method.

TranspoolConsole:
-Added a case in the commands switch for the postTripOffer command.

Input:
-Added a method that can ask the user to input a repetitionRate.

Output:
-Added the post trip offer command to the menu.
-Added a method to print the repetition rates options.
-Added a method that prints a collection of roads by src and dst.